Add bitmapbuf().
With zint_symbol.bitmap being a c_char_p, the results of
ZBarcode_Encode_and_Buffer() can't be used as a buffer
without using zint_symbol.bitmap_width and
zint_symbol.bitmap_height to determining the buffer's size.
This function returns an object of a type that includes
the buffer's size, so you don't have to import ctypes and
do it yourself.